scripts/generate_plot_data.py
#!/usr/bin/env python3
"""
Generate per-image test results and plot data files for LaTeX pgfplots.

Outputs (written to docs/report/figures/):
  - mae_vs_density.dat      (binned MAE by object density)
  - mae_vs_size.dat          (binned MAE by avg exemplar size)
  - latency_accuracy.dat     (4-point speed-accuracy trade-off)

Usage (run from code/source-code/ with CUDA available):
  python ../../scripts/generate_plot_data.py \
      --data_path ./data/FSC147/ \
      --split_file ./data/FSC147/Train_Test_Val_FSC_147.json \
      --output_dir ../../docs/report/figures

The script runs inference for 4 configurations, computes per-image MAE,
then bins results by density and object size.
"""

import argparse
import csv
import json
import math
import os
import sys
from pathlib import Path

import numpy as np
import torch
import torch.backends.cudnn as cudnn
from PIL import Image
from tqdm import tqdm

# ── Adjust path so imports work from project root ──────────────────
SCRIPT_DIR = Path(__file__).resolve().parent
CODE_DIR = SCRIPT_DIR.parent / "code" / "source-code"
sys.path.insert(0, str(CODE_DIR))

import models_mae_cross
from util.FSC147 import TransformVal  # noqa: E402

import pathlib
temp = pathlib.PosixPath
pathlib.PosixPath = pathlib.WindowsPath


# ── Configuration: checkpoint → annotation files ───────────────────
CONFIGS = [
    {
        "name": "VA-Count",
        "checkpoint": "data/checkpoint_FSC.pth",
        "anno_pos": "data/FSC147/annotation_FSC147_pos.json",
        "anno_neg": "data/FSC147/annotation_FSC147_neg.json",
        "latency": 1.4710,
    },
    {
        "name": "VA-Count + Rich Prompt",
        "checkpoint": "data/checkpoint__finetuning_dino_prompt.pth",
        "anno_pos": "data/FSC147/annotation_FSC147_pos_prompt.json",
        "anno_neg": "data/FSC147/annotation_FSC147_neg_prompt.json",
        "latency": 5.7578,
    },
    {
        "name": "VA-Count + YOLO-World",
        "checkpoint": "data/checkpoint__finetuning_yolo_noprompt.pth",
        "anno_pos": "data/FSC147/annotation_FSC147_pos_yolo.json",
        "anno_neg": "data/FSC147/annotation_FSC147_neg_yolo.json",
        "latency": 0.6006,
    },
    {
        "name": "VA-Count + YOLO-World + Rich Prompt",
        "checkpoint": "data/checkpoint__finetuning_yolo.pth",
        "anno_pos": "data/FSC147/annotation_FSC147_pos_yolo_prompt.json",
        "anno_neg": "data/FSC147/annotation_FSC147_neg_yolo_prompt.json",
        "latency": 2.4054,
    },
]


def load_model(checkpoint_path, device):
    """Load the MAE cross-attention model from a checkpoint."""
    model = models_mae_cross.mae_vit_base_patch16(norm_pix_loss=False)
    model.to(device)
    ckpt = torch.load(checkpoint_path, map_location=device)
    state = ckpt.get("model", ckpt)
    model.load_state_dict(state, strict=False)
    model.eval()
    return model


def compute_per_image_mae(model, annotations, test_imgs, data_path, device):
    """Run inference on each test image and return per-image results."""
    transform = TransformVal()
    results = []

    for img_name in tqdm(test_imgs, desc="Inference"):
        if img_name not in annotations:
            continue
        entry = annotations[img_name]
        h, w = entry["H"], entry["W"]
        points = entry.get("points", [])
        gt_count = len(points)

        # Load image
        img_path = os.path.join(data_path, "images_384_VarV2", img_name)
        if not os.path.exists(img_path):
            continue
        img = Image.open(img_path).convert("RGB")
        img_tensor = transform(img).unsqueeze(0).to(device, dtype=torch.float)

        # Load positive boxes
        boxes_raw = entry.get("box_examples_coordinates", [])
        if not boxes_raw:
            continue
        boxes = []
        for box in boxes_raw[:5]:  # up to 5 exemplars
            xs = [p[0] for p in box]
            ys = [p[1] for p in box]
            boxes.append([min(ys), min(xs), max(ys), max(xs)])
        boxes_tensor = torch.tensor(boxes, dtype=torch.float).unsqueeze(0).to(device)
        shot_num = boxes_tensor.shape[1]

        # Negative boxes
        neg_boxes_raw = entry.get("neg_box_examples_coordinates",
                                   entry.get("box_examples_coordinates", []))
        neg_boxes = []
        for box in neg_boxes_raw[:3]:
            xs = [p[0] for p in box]
            ys = [p[1] for p in box]
            neg_boxes.append([min(ys), min(xs), max(ys), max(xs)])
        neg_boxes_tensor = torch.tensor(neg_boxes, dtype=torch.float).unsqueeze(0).to(device)

        # Forward pass
        with torch.no_grad():
            with torch.amp.autocast("cuda"):
                output = model(img_tensor, boxes_tensor, shot_num)
        pred_count = torch.abs(output.sum()).item() / 60.0
        mae = abs(pred_count - gt_count)

        # Object size: average exemplar area (in original image pixels)
        areas = []
        for box in boxes_raw:
            xs = [p[0] for p in box]
            ys = [p[1] for p in box]
            bw = max(xs) - min(xs)
            bh = max(ys) - min(ys)
            areas.append(bw * bh)
        avg_size = sum(areas) / len(areas) if areas else 0

        # Density: count / megapixel
        density = gt_count / (h * w) * 1e6

        results.append({
            "image": img_name,
            "gt_count": gt_count,
            "pred_count": round(pred_count, 2),
            "mae": round(mae, 2),
            "density": round(density, 2),
            "avg_obj_size": round(avg_size, 2),
        })

    return results


def bin_data(results, key, bins):
    """Bin results by a key and compute mean MAE per bin."""
    bin_data_out = []
    for lo, hi in bins:
        items = [r for r in results if lo <= r[key] < hi]
        if items:
            mean_mae = sum(r["mae"] for r in items) / len(items)
            bin_data_out.append({
                "bin_lo": lo,
                "bin_hi": hi,
                "bin_label": f"{lo}-{hi}",
                "mean_mae": round(mean_mae, 2),
                "count": len(items),
            })
    return bin_data_out


def write_dat_file(filepath, header, rows):
    """Write a pgfplots-compatible .dat file."""
    with open(filepath, "w") as f:
        f.write("  ".join(header) + "\n")
        for row in rows:
            f.write("  ".join(str(v) for v in row) + "\n")


def main():
    parser = argparse.ArgumentParser()
    parser.add_argument("--data_path", default="./data/FSC147/")
    parser.add_argument("--split_file",
                        default="./data/FSC147/Train_Test_Val_FSC_147.json")
    parser.add_argument("--output_dir", default="../../docs/report/figures")
    parser.add_argument("--device", default="cuda")
    parser.add_argument("--config_index", type=int, default=-1,
                        help="Run only one config (0-3), or -1 for all")
    args = parser.parse_args()

    device = torch.device(args.device if torch.cuda.is_available() else "cpu")
    os.makedirs(args.output_dir, exist_ok=True)

    # Load split
    with open(args.split_file) as f:
        split = json.load(f)
    test_imgs = split["test"]
    print(f"Test images: {len(test_imgs)}")

    # Density bins (objects per megapixel)
    density_bins = [
        (0, 20), (20, 50), (50, 100), (100, 200),
        (200, 500), (500, 1000), (1000, 5000),
    ]
    # Object size bins (pixels^2)
    size_bins = [
        (0, 500), (500, 1000), (1000, 2000), (2000, 5000),
        (5000, 10000), (10000, 30000),
    ]

    configs_to_run = CONFIGS if args.config_index == -1 else [CONFIGS[args.config_index]]

    all_density_data = {}  # config_name -> binned data
    all_size_data = {}

    for cfg in configs_to_run:
        print(f"\n{'='*60}")
        print(f"Config: {cfg['name']}")
        print(f"{'='*60}")

        ckpt_path = os.path.join(os.getcwd(), cfg["checkpoint"])
        anno_path = os.path.join(os.getcwd(), cfg["anno_pos"])

        if not os.path.exists(ckpt_path):
            print(f"  WARNING: checkpoint not found: {ckpt_path}")
            continue
        if not os.path.exists(anno_path):
            print(f"  WARNING: annotation not found: {anno_path}")
            continue

        with open(anno_path) as f:
            annotations = json.load(f)

        model = load_model(ckpt_path, device)
        results = compute_per_image_mae(model, annotations, test_imgs,
                                         args.data_path, device)

        # Save per-image CSV
        csv_path = os.path.join(args.output_dir,
                                f"per_image_{cfg['name'].replace(' ', '_').replace('+', '')}.csv")
        with open(csv_path, "w", newline="") as f:
            writer = csv.DictWriter(f, fieldnames=results[0].keys())
            writer.writeheader()
            writer.writerows(results)
        print(f"  Saved {len(results)} rows to {csv_path}")

        # Aggregate MAE
        total_mae = sum(r["mae"] for r in results) / len(results)
        print(f"  Overall MAE: {total_mae:.2f}")

        # Bin by density
        density_binned = bin_data(results, "density", density_bins)
        all_density_data[cfg["name"]] = density_binned

        # Bin by object size
        size_binned = bin_data(results, "avg_obj_size", size_bins)
        all_size_data[cfg["name"]] = size_binned

        del model
        torch.cuda.empty_cache()

    # ── Write .dat files for pgfplots ──────────────────────────────
    # 1. MAE vs Density (one column per config)
    for cfg_name, binned in all_density_data.items():
        safe_name = cfg_name.replace(" ", "_").replace("+", "").lower()
        dat_path = os.path.join(args.output_dir, f"mae_vs_density_{safe_name}.dat")
        header = ["bin_mid", "mean_mae", "n"]
        rows = []
        for b in binned:
            mid = (b["bin_lo"] + b["bin_hi"]) / 2
            rows.append([mid, b["mean_mae"], b["count"]])
        write_dat_file(dat_path, header, rows)
        print(f"Wrote {dat_path}")

    # 2. MAE vs Object Size
    for cfg_name, binned in all_size_data.items():
        safe_name = cfg_name.replace(" ", "_").replace("+", "").lower()
        dat_path = os.path.join(args.output_dir, f"mae_vs_size_{safe_name}.dat")
        header = ["bin_mid", "mean_mae", "n"]
        rows = []
        for b in binned:
            mid = (b["bin_lo"] + b["bin_hi"]) / 2
            rows.append([mid, b["mean_mae"], b["count"]])
        write_dat_file(dat_path, header, rows)
        print(f"Wrote {dat_path}")

    # 3. Latency vs Accuracy (always written with hardcoded values)
    lat_acc_path = os.path.join(args.output_dir, "latency_accuracy.dat")
    header = ["latency", "mae", "label"]
    rows = [
        [c["latency"],
         sum(r["mae"] for r in compute_per_image_mae(
             load_model(os.path.join(os.getcwd(), c["checkpoint"]), device),
             json.load(open(os.path.join(os.getcwd(), c["anno_pos"]))),
             test_imgs, args.data_path, device
         )) / len(test_imgs) if False else 0,  # placeholder
         c["name"].replace(" ", "_")]
        for c in CONFIGS
    ] if False else []

    # Use known aggregate values instead
    write_dat_file(lat_acc_path,
                   ["latency", "mae", "label"],
                   [
                       [0.6006, 19.03, "YOLO-World"],
                       [1.4710, 17.99, "VA-Count"],
                       [2.4054, 17.91, "YOLO+RP"],
                       [5.7578, 17.80, "GDino+RP"],
                   ])
    print(f"Wrote {lat_acc_path}")

    print("\nDone! Copy .dat files to docs/report/figures/ and compile LaTeX.")


if __name__ == "__main__":
    main()
